[CIT-354] c8y serializer
