+++ mapper/cumulocity/c8y_translator_lib/examples/translate_to_c8yjson.rs
@@ -1,4 +1,4 @@
-use c8y_translator_lib::CumulocityJson;
+use c8y_translator_lib::json::CumulocityJson;
 
 fn single_value_translation() {
     let single_value_thin_edge_json = r#"  {
