+++ mapper/tedge_mapper/src/mapper.rs
@@ -1,4 +1,4 @@
-use c8y_translator_lib::CumulocityJson;
+use c8y_translator_lib::{json::CumulocityJson, serializer::C8yJsonSerializationError};
 use tokio::task::JoinHandle;
 use tracing::{debug, error, instrument};
 
@@ -88,7 +88,7 @@ impl Mapper {
         Ok(())
     }
 
-    fn map(input: &[u8]) -> Result<Vec<u8>, thin_edge_json::json::ThinEdgeJsonError> {
+    fn map(input: &[u8]) -> Result<Vec<u8>, C8yJsonSerializationError> {
         CumulocityJson::from_thin_edge_json(input)
     }
 }
